+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
     - Upgrade JavaFX to 21.0.1
     - Push minimal supported JDK version to 17 with a path for using it with 11
+    - Improve keyboard event handling system to support different layouts
     
 ### Bugs fixed